feat(ift_preparation): add controls for ift processing images
